orion-insight-wrangler/requirements.txt
Flask==3.0.0
google-cloud-firestore==2.16.0
PyYAML==6.0.1
gunicorn==22.0.0
google-cloud-pubsub==2.22.0
google-cloud-aiplatform==1.53.0
